# Phase - 10 API Testing/APIServer/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Testing Practice API",
    "description": "API for practicing REST Assured and API Testing",
    "version": "1.0.0"
  },
  "basePath": "/api",
  "schemes": ["http"],
  "paths": {
    "/products": {
      "get": {
        "summary": "Get all products with pagination",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "integer",
            "description": "Page number",
            "default": 1
          },
          {
            "name": "limit",
            "in": "query",
            "type": "integer",
            "description": "Items per page",
            "default": 10
          },
          {
            "name": "search",
            "in": "query",
            "type": "string",
            "description": "Keyword search"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "email": { "type": "string" },
                "password": { "type": "string" },
                "role": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login to get JWT Token",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": { "type": "string" },
                "password": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  }
}